Bootstrap 4 Documentation (printer friendly and minified) pdf
As the title suggests, this is Bootstrap 4 documentation from https://getbootstrap.com/docs/4.3/getting-started/introduction/ in quite a "minified" and printer friendly form. I was looking for it, then couldn't find anything printer friendly according to my taste and did it myself. Further explanation of relevant points to consider before tackling this document is on its first page. It's 150 pages. I hope it benefits someone